% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dat_abitibi}
\alias{dat_abitibi}
\title{Abitibi dataset}
\format{A list of 5 objects:
\itemize{
\item \code{trees} that includes identity and position of trees;
\item \code{regen2015} and \code{regen2016} regeneration and favourability data for 2015 and 2016 respectively (note that favourability values are the same for both years);
\item \code{lsdist} a list of distances: for every plot it gives the distance of all surrounding tree including in the site.
\item \code{iddist} a list of logicals of the same format as \code{lsdist} that states whether the tree is included in the buffer.
}}
\usage{
dat_abitibi
}
\description{
Trees and regeneration data for Abitibi site.
}
\keyword{datasets}
