% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatData.R
\name{formatData}
\alias{formatData}
\title{Format data.}
\usage{
formatData(fl_trees, fl_regen, fl_regen2, treesp, abbr_site,
  path_out = "output", dist_buffer = 20, export = TRUE)
}
\arguments{
\item{fl_trees}{a character string indicating path to the \code{.csv }file including the georeferenced trees.}

\item{fl_regen}{a character string indicating path to the \code{.csv} file including the plots' data for 2015 (seedling and favourability).}

\item{fl_regen2}{a character string indicating path to the \code{.csv} file including the plots' data for 2016.}

\item{treesp}{a vector of abbreviation indicating the tree species considered.}

\item{abbr_site}{a character string giving the abbreviation of the site considered.}

\item{path_out}{a character string indicating the path to write the data files.}

\item{dist_buffer}{the buffering distance (in meters).}

\item{export}{logical. If \code{TRUE}, then formatted data are written as .rds files.}
}
\value{
A list of datasets including a set of trees and their spatial coordinates, 2 #' '\code{regen} files: one for 2015 and the other for 2016 including the seedling
recruitment and the favourability of all plots and also all distances between
plots and trees and a list of trees included in a buffer.
}
\description{
Formatted to get ready for the analysis.
}
\author{
Kevin Cazelles
}
