% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getConfInt.R
\name{getConfInt}
\alias{getConfInt}
\title{Get Confidence Interval.}
\usage{
getConfInt(filename, rm_pat = "codekev12",
  basename_out = "../codekev12/resf/resf_",
  datasim = "dataReady/simu_all.Rds", quiet = TRUE)
}
\arguments{
\item{filename}{of the file to be}

\item{rm_pat}{a pattern to be removed.}

\item{basename_out}{basename of the file where results are written.}

\item{datasim}{the path of the file of simulatipon.}

\item{quiet}{logical. If \code{FALSE} then the id of the simulation is printed.}
}
\description{
A function to get confidence interval of paramters associated to one simulation.
}
\author{
Kevin Cazelles
}

