% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLikelihood.R
\name{getLikelihood}
\alias{getLikelihood}
\title{Compute the likelihood.}
\usage{
getLikelihood(pars, obs, zero_infl = FALSE, neigh = NULL,
  disp = NULL, favo = NULL, SDBH = NULL, kernel = "kern_lognormal",
  pstr = NA, ppz = NA, pscal = NA, pshap = NA, pfav = NA,
  pneigh = NA, quiet = TRUE, record = NULL)
}
\arguments{
\item{pars}{vector of parameters values.}

\item{obs}{vector of observations of recruitment for each plots of the stand.}

\item{zero_infl}{a logical. If \code{TRUE} then a zero-inflated poisson distribution is used, otherwise a poisson distribution is used.}

\item{neigh}{a vector of the size of \code{obs} describing the surrounding neighborhood. If \code{NULL} (the default setting) then neighborhood is not included in the model.}

\item{disp}{a vector of the size of \code{obs} of the surrounding trees of the focal species where seeds may be dispersed from. If \code{NULL} then dispersal is not included in the model.}

\item{favo}{a vector of the size of \code{obs} describing the composition of all parcel. If \code{NULL} then the soil favourability is not included in the model.}

\item{SDBH}{a numeric vector of standardized DBH (Diameter at breast heights).}

\item{kernel}{a character string indicating which kernel should be used (either \code{\link[diskers:kern_lognormal]{diskers::kern_lognormal()}} or \code{\link[diskers:kern_exponential_power]{diskers::kern_exponential_power()}}.}

\item{pstr}{position of parameter \code{str} in the vector of parameter values.}

\item{ppz}{position of \code{pz} parameter within the vector of parameter values.}

\item{pscal}{position of the scale parameter (of the dispersal kernel) within the vector of parameter values.}

\item{pshap}{the position of the shape parameter (of the dispersal kernel) within the vector of parameter values.}

\item{pfav}{the position of the five favourability values within the vector of parameter values.}

\item{pneigh}{position of the parameter associated to the neighborhood effect in the vector of parameter values.}

\item{quiet}{a logical. If \code{FALSE} then parameters and likelihood estimations are printed.}

\item{record}{a connection, or a character string naming the file to print to. If \code{NULL} then this step is skipped.}
}
\value{
The log-likelihood values associated to a given set of parameters.
}
\description{
This function computes the likelihood associated to a specific set of
parameters, i.e. a specific. Note that the parameters included determine the
ecological processes considered.
}
\details{
The parameter \code{pz} have a value even if it is not included in the analysis (meaning a poisson distribution is used).
}
\author{
Kevin Cazelles
}
