% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{res_bestModels}
\alias{res_bestModels}
\title{Parameters values for the best models}
\format{A data frame of 49 rows and 28 columns.}
\usage{
res_bestModels
}
\description{
Running all the simulations took hours. Tht is the reason why kept the parameters values for the best models in this dataset.
}
\author{
Kevin Cazelles
}
\keyword{datasets}
