% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieveData.R
\name{retrieveData}
\alias{retrieveData}
\title{Retrieve datasets}
\usage{
retrieveData(dir = "input")
}
\arguments{
\item{dir}{path to folder where the data will be stored.}
}
\description{
Retrieve original datasets from dryad \url{https://datadryad.org}.
}
