% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simuDesign}
\alias{simuDesign}
\title{Simulations design}
\format{A data frame of 2240 rows and 8 columns.}
\usage{
simuDesign
}
\description{
Design of the simulations as carried out in Solarik et al. 2018. Each row
represents one simulation, i.e. a combination of a site, a tree species, a
year, an age if a zero-inflated distribution was used (\code{pz} column) and
also whether favourability (\code{favo}) and neighborhood (\code{neigh})
effect were used.
}
\author{
Kevin Cazelles
}
\keyword{datasets}
