% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figKernels.R
\name{figKernels}
\alias{figKernels}
\title{Draw kernels.}
\usage{
figKernels(datares, xlim = c(0, 50), ylim = c(0, 0.015),
  colors = c("#771224", "#31679f", "#6db80f", "grey10"))
}
\arguments{
\item{datares}{the dataframe returned by the recruitement analysis.}

\item{xlim}{numeric vectors of length 2, giving the x coordinates ranges.}

\item{ylim}{numeric vectors of length 2, giving the y coordinates ranges.}

\item{colors}{vector of colors.}
}
\description{
Display the kernels associated to the best models.
}
\author{
Kevin Cazelles
}
