% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figKernels.R
\name{figKernels}
\alias{figKernels}
\title{Draw kernels.}
\usage{
figKernels(datares, age = 1, xlim = c(0, 50), ylim = c(0, 0.015),
  colors = c("#6b2934", "#4899c7", "#c8de86", "grey10"))
}
\arguments{
\item{datares}{a data frame that contains the best models.}

\item{age}{age of the seedling for which the figure is drawn (either 1 or 2).}

\item{xlim}{numeric vectors of length 2, giving the x coordinates ranges.}

\item{ylim}{numeric vectors of length 2, giving the y coordinates ranges.}

\item{colors}{vector of colors.}
}
\description{
Draws kernels of the best models.
}
\examples{
figKernels(res_bestModels)
}
\author{
Kevin Cazelles
}
