% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figKernels.R
\name{figKernels}
\alias{figKernels}
\title{Draw kernels.}
\usage{
figKernels(datares, filename = "../figs/fig1.pdf", xlim = c(0, 100),
  ylim = c(0, 0.0025))
}
\arguments{
\item{datares}{the dataframe returned by the analysis process.}

\item{filename}{a character string, giving the names of the figure to be drawn.}

\item{xlim}{numeric vectors of length 2, giving the x coordinates ranges.}

\item{ylim}{numeric vectors of length 2, giving the y coordinates ranges.}
}
\description{
A function to display the kernels associated to the best models.
}
\author{
Kevin Cazelles
}

