% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getConfInt.R
\name{getConfInt}
\alias{getConfInt}
\title{Get Confidence Intervals.}
\usage{
getConfInt(filename, simu, iter, basename_out = "inst/resf_",
  quiet = TRUE)
}
\arguments{
\item{filename}{a character string, giving the path of output file to be read.}

\item{simu}{a data frame describing the simulation.}

\item{iter}{a interger indentifying the iteration.}

\item{basename_out}{basename of the file where results are written.}

\item{quiet}{logical. If \code{FALSE} then the id of the simulation is printed.}
}
\value{
A list including the likelihood and the parameters estimates as well as the confident interval associates to the parameters.
}
\description{
This function returns confidence interval of paramters associated to a givem simulation output file.
}
\author{
Kevin Cazelles
}
