% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLikelihood.R
\name{getLikelihood}
\alias{getLikelihood}
\title{Compute the likelihood.}
\usage{
getLikelihood(pars, obs, zero_infl = FALSE, neigh = NULL, disp = NULL,
  favo = NULL, SDBH = NULL, kernel = "kern_lognormal", pstr = NA,
  ppz = NA, pscal = NA, pshap = NA, pfav = NA, pneigh = NA,
  quiet = TRUE, record = NULL)
}
\arguments{
\item{pars}{vector of parameters values, the size must}

\item{obs}{vector of observations of recruitement for each plots of the stand.}

\item{zero_infl}{logical. If \code{TRUE} then a zero-inflated poisson distribution is used, otherwose a poisson distribution is used.}

\item{neigh}{a logical. If \code{NULL} (the default setting) the is  TRUE then neighborhood}

\item{disp}{a logical. If \code{NULL} (the default setting) the is  TRUE then neighborhood}

\item{favo}{a logical. If \code{NULL} (the default setting) the is  TRUE then neighborhood}

\item{SDBH}{a numeric vector of standarized DBH (Diameter at breast heights).}

\item{kernel}{a character string indicating which kernel should be used (either \code{kern_lognormal} or \code{kern_exponential_power}). See \link[recruitR]{kernels} for further details.}

\item{pstr}{position of paraneter in thevector of parameter values.}

\item{ppz}{the position of \code{pz} parameter within the vector of parameter values.}

\item{pscal}{An integer gving the position of the scale parameter (of the dispersal kernel) witin the vector of parameter values.}

\item{pshap}{An integer gving the position of the shape parameter (of the dispersal kernel) witin the vector of parameter values.}

\item{pfav}{A vector of integer indicating the position of the five favorability values within the vector of parameter values.}

\item{pneigh}{position of the parameter associated to the neighborhood effect in the vector of parameter values.}

\item{quiet}{logial. If \code{FALSE} then parameters and likelyhood estimations are printed.}

\item{record}{a connection, or a character string naming the file to print to. If \code{NULL}, the default values, no record is done.}
}
\value{
The log-likelihood values associated to a given set of parameters.
}
\description{
This function computes the likelihood associated to a specific set of parameters.
The specification of parameters determines the ecological processes involved.
}
\details{
The parameter \code{pz} have a value even if it is not included in the analysis (meaning a poisson distrubution is used).
}
\author{
Kevin Cazelles
}
