% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getParameters.R
\name{getParameters}
\alias{getParameters}
\alias{convertTreeAbbr}
\title{Get parameters values.}
\usage{
getParameters(disp = TRUE, favo = TRUE, neigh = TRUE,
  lognormal = FALSE)

convertTreeAbbr(abbr, full_names = FALSE)
}
\arguments{
\item{disp}{a logical. If \code{TRUE}, then dispersal parameters are returned.}

\item{favo}{a logical. If \code{TRUE}, then favourability parameters are returned.}

\item{neigh}{a logical. If \code{TRUE}, then neighborhood effect parameter is returned.}

\item{lognormal}{a logical. If \code{TRUE}, then a lognormal dispersal kernel is considere.}

\item{abbr}{an abbreviation for a tree species name.}

\item{full_names}{a logical. If \code{TRUE} then scientific names of tree species are returned.}
}
\value{
A matrix including the parameters ranges and starting values.
}
\description{
A function to get ranges and starting values for the parameters associated to
a given simulation.
}
\section{Functions}{
\itemize{
\item \code{convertTreeAbbr}: Make the conversion between the two different abbreviation used for the names of tree species.
}}

\author{
Kevin Cazelles
}
