% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTableLetters.R
\name{getTableLetters}
\alias{getTableLetters}
\title{Get tables describing best models with letters.}
\usage{
getTableLetters(tmp, year)
}
\arguments{
\item{tmp}{a data.frame including results.}

\item{year}{the year for which the table is produce.}
}
\value{
A character matrix indicating the processes involved in the best model for the different seedlings.
}
\description{
Get tables describing best models with letters.
}
\author{
Kevin Cazelles
}

